requirements.txt
torch>=2.0
torchvision
numpy
pandas
matplotlib
Pillow
tqdm
requests
pyyaml
